# openapi/swagger.yaml
openapi: "3.0.3"
info:
  title: SakAI API
  description: |
    REST + WebSocket API for the SakAI ride-hailing MVP.

    ## Authentication
    All protected endpoints require a Bearer JWT **access token** in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```
    Access tokens are short-lived (60 min). Use `POST /auth/refresh` with your **refresh token**
    to silently obtain a new access token. Refresh tokens are long-lived (30 days).

    ## Idempotency
    `POST /rides` requires an `Idempotency-Key` header (client-generated UUID).
    The server caches the response for 24h — retries with the same key return the
    cached response rather than creating a duplicate ride.

    ## Error Codes
    All `4xx`/`5xx` responses include a machine-readable `code` field from the `ErrorCode` enum.
    Flutter clients should branch on `code`, not on `message`.

    ## WebSocket
    Real-time events are delivered over a persistent WebSocket connection.

    **Endpoint:** `ws://{host}/ws?token={access_token}`

    After connecting, call `GET /rides/active` to re-hydrate local state, then listen for events.
    The server filters events by identity — drivers only receive events for their assigned ride,
    passengers only for their own ride.

    All messages use the envelope format:
    ```json
    { "event": "ride.accepted", "payload": { ... } }
    ```

    See **WebSocket Event Schemas** in `components/schemas` for all payload definitions.

...
  contact:
...
servers:
...
paths:
...
  /drivers/nearby:
    get:
      tags: [Driver]
      summary: Get nearby available drivers by ride type
      description: |
        Returns a list of online drivers within the specified radius and vehicle type.
        Used by the passenger app to show available ride options with driver counts.
      operationId: getNearbyDrivers
      security:
        - BearerAuth: []
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, format: double }
        - name: lng
          in: query
          required: true
          schema: { type: number, format: double }
        - name: radius_m
          in: query
          schema: { type: number, format: double, default: 5000 }
        - name: ride_type
          in: query
          required: true
          schema:
            type: string
            enum: [motorcycle, car, tricycle]
      responses:
        "200":
          description: List of nearby drivers
          content:
            application/json:
              schema:
                type: object
                properties:
                  drivers:
                    type: array
                    items:
                      $ref: "#/components/schemas/NearbyDriver"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No drivers available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: NO_DRIVERS_AVAILABLE
                message: "No drivers of requested type within range"

  # ── Auth ──────────────────────────────────────────────────────────────────
...
  /admin/audit:
    get:
      tags: [Admin]
      summary: List audit logs
      description: Requires Superadmin role.
      operationId: adminListAudit
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: actor_id
          in: query
          required: false
          description: Filter entries to those created by this admin UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Paginated audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: Write an audit log entry
      description: |
        Records an admin action in the audit trail. Called by the frontend after
        high-impact mutations (fare changes, role edits, payout approvals).
        Requires any admin role.
      operationId: adminCreateAuditEntry
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuditEntryRequest"
      responses:
        "204":
          description: Audit entry recorded
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /admin/users/{id}/password:
    put:
      tags: [Admin]
      summary: Reset another admin's password
      description: |
        Superadmin sets a new password for another admin account.
        Distinct from `PUT /admin/auth/password` which is self-service.
        Requires Superadmin role.
      operationId: adminResetUserPassword
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "204":
          description: Password reset
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Role duplication ────────────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT access token obtained from `/auth/login` or `/auth/refresh`.
...
  responses:
    BadRequest:
      description: Invalid request payload or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: VALIDATION_ERROR
            message: "Field 'origin.lat' is required"
    Unauthorized:
      description: Missing, expired, or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: TOKEN_INVALID
            message: "Access token is missing or invalid"
    Forbidden:
      description: Authenticated user lacks permission for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: FORBIDDEN
            message: "You do not have permission to perform this action"
...
  schemas:
...
    ErrorCode:
      type: string
      description: |
        Machine-readable error code. Flutter clients should branch on this, not on `message`.
      enum:
        # Auth
        - EMAIL_ALREADY_REGISTERED
        - INVALID_CREDENTIALS
        - TOKEN_INVALID
        - TOKEN_EXPIRED
        - REFRESH_TOKEN_INVALID
        # Validation
        - VALIDATION_ERROR
        # Authorization
        - FORBIDDEN
        # Resources
        - NOT_FOUND
        - RIDE_NOT_FOUND
        - USER_NOT_FOUND
        - DOCUMENT_NOT_FOUND
        # Ride state
        - RIDE_INVALID_STATE_TRANSITION
        - PASSENGER_HAS_ACTIVE_RIDE
        - DRIVER_HAS_ACTIVE_RIDE
        - RIDE_NOT_COMPLETED
        # Matching
        - NO_DRIVERS_AVAILABLE
        # Ride type validation
        - INVALID_RIDE_TYPE
        - DRIVER_REMATCH_IN_PROGRESS
        - CANCELLATION_FEE_APPLIED
        # Payments
        - PAYMENT_FAILED
        - INVALID_PAYMENT_TOKEN
        - DUPLICATE_PAYMENT
        - UNPAID_RIDE_BLOCKED
        # Payment Methods
        - PAYMENT_METHOD_UNSUPPORTED
        - PAYMENT_METHOD_DUPLICATE
        - PAYMENT_GATEWAY_ERROR
        - PAYMENT_METHOD_NOT_FOUND
        - PAYMENT_METHOD_LAST_METHOD
        # Tips
        - INVALID_TIP_AMOUNT
        - TIP_ALREADY_ADDED
        # Ratings
        - INVALID_RATING
        - FEEDBACK_TOO_LONG
        - ALREADY_RATED
        # File uploads
        - FILE_TOO_LARGE
        - INVALID_FILE_FORMAT
        - INVALID_DOCUMENT_TYPE
        # Operational
        - RATE_LIMIT_EXCEEDED
        - INTERNAL_SERVER_ERROR
        # Proximity
        - DRIVER_TOO_FAR
        - DRIVER_TOO_FAR_FROM_DESTINATION
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Human-readable error description (for logging/display fallback)
          example: "An account with this email already exists"
        details:
          type: object
          description: Optional field-level validation details
          additionalProperties: true
          example:
            field: "origin.lat"
            reason: "must be between -90 and 90"

    # -- Common ---------------------------------------------------------------
    LatLng:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 14.5995
        lng:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: 120.9842

    # -- System ---------------------------------------------------------------
...
    AuditLogResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
    AuditLog:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        actor_id:
          type: string
        ip_address:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        before_state:
          type: object
          nullable: true
          additionalProperties: true
        after_state:
          type: object
          nullable: true
          additionalProperties: true
        reason:
          type: string
          nullable: true
...
    ResetPasswordRequest:
      type: object
      required: [new_password]
      properties:
        new_password:
          type: string
          format: password
          minLength: 8
    CreateAuditEntryRequest:
      type: object
      required: [resource_type, resource_id, action]
      properties:
        resource_type:
          type: string
          example: fare_config
        resource_id:
          type: string
        action:
          type: string
          enum: [create, update, delete, approve, reject, login, logout]
        before_state:
          type: object
          nullable: true
          additionalProperties: true
        after_state:
          type: object
          nullable: true
          additionalProperties: true
        reason:
          type: string
          nullable: true
...
    NearbyDriver:
      type: object
      required: [id, name, location, vehicle_type]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        vehicle_make:
          type: string
        vehicle_model:
          type: string
        vehicle_plate:
          type: string
        vehicle_type:
          type: string
          enum: [motorcycle, car, tricycle]
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          nullable: true
        distance_m:
          type: number
          format: double
          nullable: true
        location:
          $ref: "#/components/schemas/LatLng"
        heading:
          type: number
          format: double
          nullable: true
          description: Compass bearing in degrees
...